raya-bot/requirements.txt
discord.py
requests
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib